Refactor pan/fling support
